Lift bookmarks code out of RepoUtils. Make test ignoring of special folders explicit.
