+++ test/src/org/mozilla/android/sync/test/helpers/ExpectGuidsSinceDelegate.java
@@ -7,6 +7,8 @@ import static junit.framework.Assert.assertEquals;
 import static junit.framework.Assert.assertFalse;
 
 import java.util.Arrays;
+import java.util.HashSet;
+import java.util.Set;
 
 import android.util.Log;
 
@@ -14,6 +16,7 @@ import junit.framework.AssertionFailedError;
 
 public class ExpectGuidsSinceDelegate extends DefaultGuidsSinceDelegate {
   private String[] expected;
+  public Set<String> ignore = new HashSet<String>();
 
   public ExpectGuidsSinceDelegate(String[] guids) {
     expected = guids;
@@ -25,11 +28,14 @@ public class ExpectGuidsSinceDelegate extends DefaultGuidsSinceDelegate {
   public void onGuidsSinceSucceeded(String[] guids) {
     AssertionFailedError err = null;
     try {
-      assertEquals(this.expected.length, guids.length);
-
-      for (String string : guids) {
-        assertFalse(-1 == Arrays.binarySearch(this.expected, string));
+      int notIgnored = 0;
+      for (String guid : guids) {
+        if (!ignore.contains(guid)) {
+          notIgnored++;
+          assertFalse(-1 == Arrays.binarySearch(this.expected, guid));
+        }
       }
+      assertEquals(this.expected.length, notIgnored);
     } catch (AssertionFailedError e) {
       err = e;
     }
