+++ test/src/org/mozilla/android/sync/test/helpers/ExpectNoGUIDsSinceDelegate.java
@@ -5,23 +5,29 @@ package org.mozilla.android.sync.test.helpers;
 
 import static junit.framework.Assert.assertEquals;
 
-import org.mozilla.gecko.sync.repositories.android.RepoUtils;
+import java.util.HashSet;
+import java.util.Set;
+
+import android.util.Log;
 
 import junit.framework.AssertionFailedError;
 
 public class ExpectNoGUIDsSinceDelegate extends DefaultGuidsSinceDelegate {
   
+  public Set<String> ignore = new HashSet<String>();
+
   @Override
   public void onGuidsSinceSucceeded(String[] guids) {
     AssertionFailedError err = null;
     try {
-      boolean bookmarks = false;
+      int nonIgnored = 0;
       for (int i = 0; i < guids.length; i++) {
-        if(RepoUtils.SPECIAL_GUIDS_MAP.containsKey(guids[i])) {
-          bookmarks = true;
+        Log.i("BOOM", "Got " + guids[i]);
+        if (!ignore.contains(guids[i])) {
+          nonIgnored++;
         }
       }
-      assertEquals(0, bookmarks ? guids.length - 5 : guids.length);
+      assertEquals(0, nonIgnored);
     } catch (AssertionFailedError e) {
       err = e;
     }
